Factor out javascript logic into multiple files
