+++ web/index.html
@@ -22,6 +22,6 @@
     </div>
 
     <script src="https://d3js.org/d3.v7.min.js"></script>
-    <script src="/main.js"></script>
+    <script type="module" src="/main.js"></script>
 </body>
 </html>
